// src/models/roleEnum.ts
/**
 * Generated by orval
 * Do not edit manually.
 */

/**
 * * `DSTAX_ADMIN` - DSTax_Admin
 * `DSTAX_PREPARER` - DSTax_Preparer
 * `CLIENT_ADMIN` - Client_Admin
 * `CLIENT_STAFF` - Client_Staff
 */
export type RoleEnum = (typeof RoleEnum)[keyof typeof RoleEnum]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RoleEnum = {
  DSTAX_ADMIN: 'DSTAX_ADMIN',
  DSTAX_PREPARER: 'DSTAX_PREPARER',
  CLIENT_ADMIN: 'CLIENT_ADMIN',
  CLIENT_STAFF: 'CLIENT_STAFF',
} as const